another way to implement retag. maybe cleaner?
This is slower than the version where the function is fused, but I think it
involves the same number of copies. The code is simpler, so I'm ok for now with
the 50 ms hit, since whatever is causing this slowdown is probably slowing down
all other operations too.

task demo timing:
retag expression: 171 ms.
Computing stats: 36 ms.